+++ managers/date_greater_than_filter.py
@@ -5,16 +5,20 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.future import select
 from models.date_greater_than_filter import DateGreaterThanFilter
 from models.serialization_schema.date_greater_than_filter import DateGreaterThanFilterSchema
+class DateGreaterThanFilterNotFoundError(Exception):
+    pass
 class DateGreaterThanFilterManager:
     def __init__(self, session: AsyncSession):
         self.session = session
-    def build(self, **kwargs) -> DateGreaterThanFilter:
+    async def build(self, **kwargs) -> DateGreaterThanFilter:
         return DateGreaterThanFilter(**kwargs)
     async def add(self, date_greater_than_filter: DateGreaterThanFilter) -> DateGreaterThanFilter:
         self.session.add(date_greater_than_filter)
         await self.session.commit()
         return date_greater_than_filter
     async def get_by_id(self, date_greater_than_filter_id: int) -> Optional[DateGreaterThanFilter]:
+        if not isinstance(date_greater_than_filter_id, int):
+            raise TypeError(f"The date_greater_than_filter_id must be an integer, got {type(date_greater_than_filter_id)} instead.")
         result = await self.session.execute(select(DateGreaterThanFilter).filter(DateGreaterThanFilter.date_greater_than_filter_id == date_greater_than_filter_id))
         return result.scalars().first()
     async def get_by_code(self, code: uuid.UUID) -> Optional[DateGreaterThanFilter]:
@@ -26,12 +30,14 @@ class DateGreaterThanFilterManager:
                 setattr(date_greater_than_filter, key, value)
             await self.session.commit()
         return date_greater_than_filter
-    async def delete(self, date_greater_than_filter_id: int) -> Optional[DateGreaterThanFilter]:
+    async def delete(self, date_greater_than_filter_id: int):
+        if not isinstance(date_greater_than_filter_id, int):
+            raise TypeError(f"The date_greater_than_filter_id must be an integer, got {type(date_greater_than_filter_id)} instead.")
         date_greater_than_filter = await self.get_by_id(date_greater_than_filter_id)
-        if date_greater_than_filter:
-            self.session.delete(date_greater_than_filter)
-            await self.session.commit()
-        return date_greater_than_filter
+        if not date_greater_than_filter:
+            raise DateGreaterThanFilterNotFoundError(f"DateGreaterThanFilter with ID {date_greater_than_filter_id} not found!")
+        await self.session.delete(date_greater_than_filter)
+        await self.session.commit()
     async def get_list(self) -> List[DateGreaterThanFilter]:
         result = await self.session.execute(select(DateGreaterThanFilter))
         return result.scalars().all()
@@ -42,6 +48,13 @@ class DateGreaterThanFilterManager:
         schema = DateGreaterThanFilterSchema()
         date_greater_than_filter_data = schema.dump(date_greater_than_filter)
         return json.dumps(date_greater_than_filter_data)
+    def to_dict(self, date_greater_than_filter:DateGreaterThanFilter) -> dict:
+        """
+        Serialize the DateGreaterThanFilter object to a JSON string using the DateGreaterThanFilterSchema.
+        """
+        schema = DateGreaterThanFilterSchema()
+        date_greater_than_filter_data = schema.dump(date_greater_than_filter)
+        return date_greater_than_filter_data
     def from_json(self, json_str: str) -> DateGreaterThanFilter:
         """
         Deserialize a JSON string into a DateGreaterThanFilter object using the DateGreaterThanFilterSchema.
@@ -51,9 +64,8 @@ class DateGreaterThanFilterManager:
         date_greater_than_filter_dict = schema.load(data)
         new_date_greater_than_filter = DateGreaterThanFilter(**date_greater_than_filter_dict)
         return new_date_greater_than_filter
-    async def add_bulk(self, date_greater_than_filters_data: List[Dict]) -> List[DateGreaterThanFilter]:
+    async def add_bulk(self, date_greater_than_filters: List[DateGreaterThanFilter]) -> List[DateGreaterThanFilter]:
         """Add multiple date_greater_than_filters at once."""
-        date_greater_than_filters = [DateGreaterThanFilter(**data) for data in date_greater_than_filters_data]
         self.session.add_all(date_greater_than_filters)
         await self.session.commit()
         return date_greater_than_filters
@@ -62,11 +74,13 @@ class DateGreaterThanFilterManager:
         updated_date_greater_than_filters = []
         for update in date_greater_than_filter_updates:
             date_greater_than_filter_id = update.get("date_greater_than_filter_id")
+            if not isinstance(date_greater_than_filter_id, int):
+                raise TypeError(f"The date_greater_than_filter_id must be an integer, got {type(date_greater_than_filter_id)} instead.")
             if not date_greater_than_filter_id:
                 continue
             date_greater_than_filter = await self.get_by_id(date_greater_than_filter_id)
             if not date_greater_than_filter:
-                continue
+                raise DateGreaterThanFilterNotFoundError(f"DateGreaterThanFilter with ID {date_greater_than_filter_id} not found!")
             for key, value in update.items():
                 if key != "date_greater_than_filter_id":
                     setattr(date_greater_than_filter, key, value)
@@ -76,9 +90,13 @@ class DateGreaterThanFilterManager:
     async def delete_bulk(self, date_greater_than_filter_ids: List[int]) -> bool:
         """Delete multiple date_greater_than_filters by their IDs."""
         for date_greater_than_filter_id in date_greater_than_filter_ids:
+            if not isinstance(date_greater_than_filter_id, int):
+                raise TypeError(f"The date_greater_than_filter_id must be an integer, got {type(date_greater_than_filter_id)} instead.")
             date_greater_than_filter = await self.get_by_id(date_greater_than_filter_id)
+            if not date_greater_than_filter:
+                raise DateGreaterThanFilterNotFoundError(f"DateGreaterThanFilter with ID {date_greater_than_filter_id} not found!")
             if date_greater_than_filter:
-                self.session.delete(date_greater_than_filter)
+                await self.session.delete(date_greater_than_filter)
         await self.session.commit()
         return True
     async def count(self) -> int:
@@ -94,14 +112,18 @@ class DateGreaterThanFilterManager:
         return result.scalars().all()
     async def refresh(self, date_greater_than_filter: DateGreaterThanFilter) -> DateGreaterThanFilter:
         """Refresh the state of a given date_greater_than_filter instance from the database."""
-        self.session.refresh(date_greater_than_filter)
+        await self.session.refresh(date_greater_than_filter)
         return date_greater_than_filter
     async def exists(self, date_greater_than_filter_id: int) -> bool:
         """Check if a date_greater_than_filter with the given ID exists."""
+        if not isinstance(date_greater_than_filter_id, int):
+            raise TypeError(f"The date_greater_than_filter_id must be an integer, got {type(date_greater_than_filter_id)} instead.")
         date_greater_than_filter = await self.get_by_id(date_greater_than_filter_id)
         return bool(date_greater_than_filter)
 
     async def get_by_pac_id(self, pac_id: int): # PacID
+        if not isinstance(pac_id, int):
+            raise TypeError(f"The date_greater_than_filter_id must be an integer, got {type(pac_id)} instead.")
         result = await self.session.execute(select(DateGreaterThanFilter).filter(DateGreaterThanFilter.pac_id == pac_id))
         return result.scalars().all()
